\name{decisivatoR-package}
\alias{decisivatoR-package}
\alias{decisivatoR}
\docType{package}

\title{

  decisivatoR: an R infrastructure package that addresses the problem of phylogenetic decisiveness.

}
\description{

 An R infrastructure package that was inspired from work of Steel, Sanderson and Fischer. decisivatoR is a practical, convenient, user-oriented application to address the problem of phylogenetic decisiveness for provided data set.

}
\details{

  See \code{packageDescription('decisivatoR')}

}
\author{
Ilya Y. Zhbannikov <i.zhbannikov@gmail.com>, James A. Foster <foster@uidaho.edu>
}
\maintainer{
Ilya Y. Zhbannikov <i.zhbannikov@gmail.com>
}
\references{

Mike Steel, Michael J Sanderson, \emph{Characterizing phylogenetically decisive taxon coverage},
	Appl Math Letters,23:82-86,2010

Michael J Sanderson, Michelle M McMahon, Mike Steel, \emph{Phylogenomics with incomplete taxon coverage: the limits to inference},
	BMC evolutionary biology,2010
	
Mareike Fischer, \emph{Perfect taxon sampling and phylogenetically decisive taxon coverage},
	arXiv preprint arXiv:1206.3472, June, 2012

Mareike Fischer, \emph{Mathematical aspects of phylogenetic groves}, Annals of Combinatorics, 2013

Semple C, SteelM, \emph{Phylogenetics},
	Oxford University Press, 2003
	
}
\keyword{ package }
