\name{decisivatoR-package}
\alias{decisivatoR-package}
\alias{decisivatoR}
\docType{package}

\title{

  decisivatoR: an R infrastructure package that addresses the problem of phylogenetic decisiveness.

}
\description{

 An R infrastructure package that was inspired from work of Steel, Sanderson and Fischer. decisivatoR is a practical, convenient, user-oriented application to address the problem of phylogenetic decisiveness for provided data set.

}
\details{

  See \code{packageDescription('decisivatoR')}

}
\author{
Ilya Y. Zhbannikov
Maintainer: Ilya Y. Zhbannikov <i.zhbannikov@gmail.com>
}
\references{

}
\keyword{ package }
